gallery: Create , multiple file upload success, Delete gallery with unused images. Delete images. Next -> sending extra information as gallery title and image comment

+++ Application/Galleries/Delete.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Application.Core;
@@ -8,11 +9,12 @@ using Persistence;
 
 namespace Application.Galleries
 {
-    public class Delete
+    public class DeleteEventoGallery
     {
         public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
+            public Guid EventoId { get; set; }
         }
 
         public class Handler : IRequestHandler<Command, Result<Unit>>
@@ -26,15 +28,37 @@ namespace Application.Galleries
             public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var gallery = await _context.Galleries.FindAsync(request.Id);
-                var images = gallery.Images;
-                foreach (var item in images)
-                {
-                    var filename = "C:\\workspace\\AjedrezLanzarote\\client-app\\public\\assets\\galleryImages\\"+item.Filename;
-                    var thumbfile = "C:\\workspace\\AjedrezLanzarote\\client-app\\public\\assets\\galleryImages\\"+item.Thumbnail;
-                    System.IO.File.Delete(filename);
-                    System.IO.File.Delete(thumbfile);
+                var galleryImages = await _context.GalleryImages.Where(x => x.GalleryId == gallery.Id).ToListAsync();
+                
+                var eventoGallery = await _context.GalleryEventos.FindAsync(request.Id, request.EventoId);
+                
+                
+                var relationsEvento = _context.GalleryEventos.AnyAsync(x => x.GalleryId == gallery.Id && x.EventoId != request.EventoId).Result;
+                if (relationsEvento) {
+                    
+                    _context.Remove(eventoGallery);
+                    var resultFirst = await _context.SaveChangesAsync() > 0;
+                    if (!resultFirst) return Result<Unit>.Failure("Fallo al eliminar una galeria");
+                    return Result<Unit>.Success(Unit.Value);
                 }
-                //if (evento == null) return null;
+                
+                foreach (var item in galleryImages)
+                {
+                    var imageObject = _context.Images.Where(x => x.Id == item.ImageId).FirstOrDefaultAsync();
+                    
+                    var relatedimages = await _context.GalleryImages.AnyAsync(x => x.ImageId == imageObject.Result.Id && x.GalleryId != gallery.Id);
+                    if (relatedimages == false)
+                    {
+                        var filename = "C:\\workspace\\AjedrezLanzarote\\client-app\\public\\assets\\galleryImages\\"+imageObject.Result.Filename;
+                        var thumbfile = "C:\\workspace\\AjedrezLanzarote\\client-app\\public\\"+imageObject.Result.Thumbnail;
+                        System.IO.File.Delete(filename);
+                        System.IO.File.Delete(thumbfile);
+                        _context.Remove(imageObject.Result);
+                    }
+                    
+                }
+                _context.Remove(eventoGallery);
+
                 _context.Remove(gallery);
 
                 var result = await _context.SaveChangesAsync() > 0;
